LPS-105162 move 1_0_0 to 1_1_0
