+++ modules/dxp/apps/saml/saml-persistence-service/src/main/java/com/liferay/saml/persistence/internal/upgrade/v1_1_0/UpgradeSamlIdpSpSession.java
@@ -12,10 +12,10 @@
  *
  */
 
-package com.liferay.saml.persistence.internal.upgrade.v1_0_0;
+package com.liferay.saml.persistence.internal.upgrade.v1_1_0;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.saml.persistence.internal.upgrade.v1_0_0.util.SamlSpAuthRequestTable;
+import com.liferay.saml.persistence.internal.upgrade.v1_1_0.util.SamlIdpSpSessionTable;
 
 import java.sql.SQLException;
 
@@ -23,21 +23,27 @@ import java.sql.SQLException;
  * @author Mika Koivisto
  * @author Brian Wing Shun Chan
  */
-public class UpgradeSamlSpAuthRequest extends UpgradeProcess {
+public class UpgradeSamlIdpSpSession extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
 		try {
 			runSQL(
-				"alter_column_type SamlSpAuthRequest samlIdpEntityId " +
+				"alter_column_type SamlIdpSpSession samlSpEntityId " +
+					"VARCHAR(1024) null");
+			runSQL(
+				"alter_column_type SamlIdpSpSession nameIdFormat " +
+					"VARCHAR(1024) null");
+			runSQL(
+				"alter_column_type SamlIdpSpSession nameIdValue " +
 					"VARCHAR(1024) null");
 		}
 		catch (SQLException sqlException) {
 			upgradeTable(
-				SamlSpAuthRequestTable.TABLE_NAME,
-				SamlSpAuthRequestTable.TABLE_COLUMNS,
-				SamlSpAuthRequestTable.TABLE_SQL_CREATE,
-				SamlSpAuthRequestTable.TABLE_SQL_ADD_INDEXES);
+				SamlIdpSpSessionTable.TABLE_NAME,
+				SamlIdpSpSessionTable.TABLE_COLUMNS,
+				SamlIdpSpSessionTable.TABLE_SQL_CREATE,
+				SamlIdpSpSessionTable.TABLE_SQL_ADD_INDEXES);
 		}
 	}
