+++ modules/dxp/apps/saml/saml-persistence-service/src/main/java/com/liferay/saml/persistence/internal/upgrade/v1_1_0/UpgradeSamlSpAuthRequest.java
@@ -12,10 +12,10 @@
  *
  */
 
-package com.liferay.saml.persistence.internal.upgrade.v1_0_0;
+package com.liferay.saml.persistence.internal.upgrade.v1_1_0;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.saml.persistence.internal.upgrade.v1_0_0.util.SamlSpMessageTable;
+import com.liferay.saml.persistence.internal.upgrade.v1_1_0.util.SamlSpAuthRequestTable;
 
 import java.sql.SQLException;
 
@@ -23,20 +23,21 @@ import java.sql.SQLException;
  * @author Mika Koivisto
  * @author Brian Wing Shun Chan
  */
-public class UpgradeSamlSpMessage extends UpgradeProcess {
+public class UpgradeSamlSpAuthRequest extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
 		try {
 			runSQL(
-				"alter_column_type SamlSpMessage samlIdpEntityId " +
+				"alter_column_type SamlSpAuthRequest samlIdpEntityId " +
 					"VARCHAR(1024) null");
 		}
 		catch (SQLException sqlException) {
 			upgradeTable(
-				SamlSpMessageTable.TABLE_NAME, SamlSpMessageTable.TABLE_COLUMNS,
-				SamlSpMessageTable.TABLE_SQL_CREATE,
-				SamlSpMessageTable.TABLE_SQL_ADD_INDEXES);
+				SamlSpAuthRequestTable.TABLE_NAME,
+				SamlSpAuthRequestTable.TABLE_COLUMNS,
+				SamlSpAuthRequestTable.TABLE_SQL_CREATE,
+				SamlSpAuthRequestTable.TABLE_SQL_ADD_INDEXES);
 		}
 	}
