+++ modules/dxp/apps/saml/saml-persistence-service/src/main/java/com/liferay/saml/persistence/internal/upgrade/v1_1_0/UpgradeSamlSpMessage.java
@@ -12,10 +12,10 @@
  *
  */
 
-package com.liferay.saml.persistence.internal.upgrade.v1_0_0;
+package com.liferay.saml.persistence.internal.upgrade.v1_1_0;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.saml.persistence.internal.upgrade.v1_0_0.util.SamlSpSessionTable;
+import com.liferay.saml.persistence.internal.upgrade.v1_1_0.util.SamlSpMessageTable;
 
 import java.sql.SQLException;
 
@@ -23,23 +23,20 @@ import java.sql.SQLException;
  * @author Mika Koivisto
  * @author Brian Wing Shun Chan
  */
-public class UpgradeSamlSpSession extends UpgradeProcess {
+public class UpgradeSamlSpMessage extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
 		try {
 			runSQL(
-				"alter_column_type SamlSpSession nameIdFormat VARCHAR(1024) " +
-					"null");
-			runSQL(
-				"alter_column_type SamlSpSession nameIdValue VARCHAR(1024) " +
-					"null");
+				"alter_column_type SamlSpMessage samlIdpEntityId " +
+					"VARCHAR(1024) null");
 		}
 		catch (SQLException sqlException) {
 			upgradeTable(
-				SamlSpSessionTable.TABLE_NAME, SamlSpSessionTable.TABLE_COLUMNS,
-				SamlSpSessionTable.TABLE_SQL_CREATE,
-				SamlSpSessionTable.TABLE_SQL_ADD_INDEXES);
+				SamlSpMessageTable.TABLE_NAME, SamlSpMessageTable.TABLE_COLUMNS,
+				SamlSpMessageTable.TABLE_SQL_CREATE,
+				SamlSpMessageTable.TABLE_SQL_ADD_INDEXES);
 		}
 	}
