+++ modules/dxp/apps/saml/saml-persistence-service/src/main/java/com/liferay/saml/persistence/internal/upgrade/v1_1_0/UpgradeSamlSpSession.java
@@ -12,10 +12,10 @@
  *
  */
 
-package com.liferay.saml.persistence.internal.upgrade.v1_0_0;
+package com.liferay.saml.persistence.internal.upgrade.v1_1_0;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.saml.persistence.internal.upgrade.v1_0_0.util.SamlIdpSpSessionTable;
+import com.liferay.saml.persistence.internal.upgrade.v1_1_0.util.SamlSpSessionTable;
 
 import java.sql.SQLException;
 
@@ -23,27 +23,23 @@ import java.sql.SQLException;
  * @author Mika Koivisto
  * @author Brian Wing Shun Chan
  */
-public class UpgradeSamlIdpSpSession extends UpgradeProcess {
+public class UpgradeSamlSpSession extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
 		try {
 			runSQL(
-				"alter_column_type SamlIdpSpSession samlSpEntityId " +
-					"VARCHAR(1024) null");
+				"alter_column_type SamlSpSession nameIdFormat VARCHAR(1024) " +
+					"null");
 			runSQL(
-				"alter_column_type SamlIdpSpSession nameIdFormat " +
-					"VARCHAR(1024) null");
-			runSQL(
-				"alter_column_type SamlIdpSpSession nameIdValue " +
-					"VARCHAR(1024) null");
+				"alter_column_type SamlSpSession nameIdValue VARCHAR(1024) " +
+					"null");
 		}
 		catch (SQLException sqlException) {
 			upgradeTable(
-				SamlIdpSpSessionTable.TABLE_NAME,
-				SamlIdpSpSessionTable.TABLE_COLUMNS,
-				SamlIdpSpSessionTable.TABLE_SQL_CREATE,
-				SamlIdpSpSessionTable.TABLE_SQL_ADD_INDEXES);
+				SamlSpSessionTable.TABLE_NAME, SamlSpSessionTable.TABLE_COLUMNS,
+				SamlSpSessionTable.TABLE_SQL_CREATE,
+				SamlSpSessionTable.TABLE_SQL_ADD_INDEXES);
 		}
 	}
